+++ frontend/src/components/Footer/index.js
