+++ frontend/src/index.js
@@ -1,8 +1,8 @@
 import { render } from "react-dom"
 import App from "./App"
 import { BrowserRouter, Routes, Route } from "react-router-dom"
-import AddProduct from "./components/pages/AddProduct"
-import AddCategory from "./components/pages/AddCategory"
+import AddProduct from "./pages/AddProduct"
+import AddCategory from "./pages/AddCategory"
 import { Slide, ToastContainer } from "react-toastify"
 import "react-toastify/dist/ReactToastify.css"
